Corrigiendo el tipo de dato del id

+++ BackEnd/elCaserio/src/main/java/com/elcaserio/repository/IAlergenoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IAlergenoRepository extends JpaRepository <Alergeno, Integer> {
+public interface IAlergenoRepository extends JpaRepository <Alergeno, Long> {
 }
